requirements.txt
streamlit==1.53.1
groq==1.0.0
pymupdf==1.26.7
python-dotenv==1.2.1
apify-client==2.4.1
requests==2.32.5
pandas==2.3.3
numpy==2.4.1
pillow==12.1.0
altair==6.0.0
plotly==5.24.1
matplotlib==3.9.3
reportlab==4.2.5
fpdf2==2.8.1